Generative AI/basic/requirement.txt
langchain 
langchain-core

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub

python-dotenv

numpy
scikit-learn